tools/tafsir_gui/tests/requirements.txt
pytest-playwright>=0.7.0
playwright>=1.4.0
